Refactor: move location models to data.models
