+++ app/src/main/java/fr/free/nrw/commons/Utils.java
@@ -17,7 +17,6 @@ import androidx.browser.customtabs.CustomTabColorSchemeParams;
 import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.core.content.ContextCompat;
 
-import fr.free.nrw.commons.kvstore.JsonKvStore;
 import java.util.Date;
 import org.wikipedia.dataclient.WikiSite;
 import org.wikipedia.page.PageTitle;
@@ -25,13 +24,12 @@ import org.wikipedia.page.PageTitle;
 import java.util.Locale;
 import java.util.regex.Pattern;
 
-import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.data.models.location.LatLng;
 import fr.free.nrw.commons.settings.Prefs;
 import fr.free.nrw.commons.utils.ViewUtil;
 import timber.log.Timber;
 
 import static android.widget.Toast.LENGTH_SHORT;
-import static fr.free.nrw.commons.campaigns.CampaignView.CAMPAIGNS_DEFAULT_PREFERENCE;
 
 public class Utils {
