+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsFragment.java
@@ -31,7 +31,6 @@ import fr.free.nrw.commons.auth.SessionManager;
 import fr.free.nrw.commons.notification.Notification;
 import fr.free.nrw.commons.notification.NotificationController;
 import fr.free.nrw.commons.theme.BaseActivity;
-import java.text.ParseException;
 import java.util.Date;
 import java.util.List;
 import javax.inject.Inject;
@@ -49,7 +48,7 @@ import fr.free.nrw.commons.contributions.ContributionsListFragment.Callback;
 import fr.free.nrw.commons.contributions.MainActivity.ActiveFragment;
 import fr.free.nrw.commons.di.CommonsDaggerSupportFragment;
 import fr.free.nrw.commons.kvstore.JsonKvStore;
-import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.data.models.location.LatLng;
 import fr.free.nrw.commons.location.LocationServiceManager;
 import fr.free.nrw.commons.location.LocationUpdateListener;
 import fr.free.nrw.commons.media.MediaDetailPagerFragment;
