+++ app/src/main/java/fr/free/nrw/commons/location/LocationServiceManager.java
@@ -9,8 +9,8 @@ import android.location.Location;
 import android.location.LocationListener;
 import android.location.LocationManager;
 import android.os.Bundle;
-import android.util.Log;
 import androidx.core.app.ActivityCompat;
+import fr.free.nrw.commons.data.models.location.LatLng;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
