+++ app/src/main/java/fr/free/nrw/commons/location/LocationUpdateListener.java
@@ -1,5 +1,7 @@
 package fr.free.nrw.commons.location;
 
+import fr.free.nrw.commons.data.models.location.LatLng;
+
 public interface LocationUpdateListener {
     void onLocationChangedSignificantly(LatLng latLng); // Will be used to update all nearby markers on the map
     void onLocationChangedSlightly(LatLng latLng); // Will be used to track users motion
