+++ app/src/main/java/fr/free/nrw/commons/nearby/NearbyPlaces.java
@@ -8,7 +8,7 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.data.models.location.LatLng;
 import fr.free.nrw.commons.mwapi.OkHttpJsonApiClient;
 import timber.log.Timber;
