+++ app/src/main/java/fr/free/nrw/commons/upload/UploadView.java
@@ -7,7 +7,7 @@ import androidx.annotation.IntDef;
 import java.lang.annotation.Retention;
 import java.util.List;
 
-import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.data.models.location.LatLng;
 
 import static java.lang.annotation.RetentionPolicy.SOURCE;
