+++ app/src/main/java/fr/free/nrw/commons/upload/WikidataPlace.kt
@@ -1,7 +1,7 @@
 package fr.free.nrw.commons.upload
 
 import android.os.Parcelable
-import fr.free.nrw.commons.location.LatLng
+import fr.free.nrw.commons.data.models.location.LatLng
 import fr.free.nrw.commons.nearby.Place
 import kotlinx.android.parcel.Parcelize
