+++ app/src/main/java/fr/free/nrw/commons/utils/PlaceUtils.java
@@ -3,7 +3,7 @@ package fr.free.nrw.commons.utils;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.data.models.location.LatLng;
 
 public class PlaceUtils {
