+++ app/src/test/kotlin/fr/free/nrw/commons/nearby/NearbyPlacesTest.kt
@@ -4,7 +4,7 @@ import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
-import fr.free.nrw.commons.location.LatLng
+import fr.free.nrw.commons.data.models.location.LatLng
 import fr.free.nrw.commons.mwapi.OkHttpJsonApiClient
 import org.junit.Before
 import org.junit.Test
